Add files via upload
Your assignment is as follows:
You need to come up with an algorithm that works which can take the following PDF and extract key information from the same for an investor looking to evaluate the company. The investor looks at key elements such as future growth prospects, key changes in the business, key triggers, important information that might have a material effect on next year's earnings and growth. 


In this assignment you are free to use the GPT algorithm and API's of your choice. You are free to use any python libraries and packages to achieve this task. The code should be well documented and cleanly written. 

Length of code is not important, its quality is. Also the quality of the output is important. The code should work across other documents of the same kind and should preferably not have any hard coded logic within it. 

You will need to submit your code as a Google collab link (Please ensure the Google collab link is visible to public for anyone who has the link).


Please use the below link to access the document that needs to be summarized:

https://drive.google.com/file/d/1jguXFqGgYkbeF5X9jFJ4ts50m-VJ1rCu/view?usp=sharing